(#63) Boxes: move to the Shared assembly
