+++ src/WpfMath.Shared/Boxes/GlueBox.cs
