+++ src/WpfMath.Shared/Boxes/HorizontalBox.cs
