+++ src/WpfMath.Shared/Boxes/OverBar.cs
