+++ src/WpfMath.Shared/Boxes/StrutBox.cs
