+++ src/WpfMath.Shared/Boxes/VerticalBox.cs
