+++ src/WpfMath/TexRenderer.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Windows;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using WpfMath.Boxes;
